fixed the index script

+++ apply_index_local.py
@@ -5,7 +5,7 @@ import shutil
 import settings
 
 from mii_common import tools
-
+from time import sleep
 
 def apply_index(path, json_file_name):
     json_file_path = os.path.join(path, json_file_name)
@@ -17,6 +17,13 @@ def apply_index(path, json_file_name):
     index_path = os.path.join(path, 'Movies', 'Index')
     if os.path.exists(index_path):
         shutil.rmtree(index_path)
+    retry = 0
+    while os.path.exists(index_path):
+	retry += 1
+        print 'Sleeping until deleted'
+	sleep(10)
+	if retry == 5:
+	    return
     current_path_root = tools.make_dir(index_path)
     tools.dict_apply(current_path_root, dict_index, symlink_method=os.symlink)
     os.remove(json_file_path)
